move static functions in core, add files for vectors and statistical test static functions
